PyRAF script to clean telluric lines from science data
First, a continuum is fit (using the task CONTINUUM) to a spectrum collected from a standard star, and the spectrum is normalized. Then, SPLOT is used interactively to clean the normalized spectrum of obnoxious spikes at the edges, as well as the H-Br gamma peak seen at ~21660 Ang. Then, the TELLURIC task is used to clean the OBJ data.